Created font.h and font_renderer_default backend
