+++ engine/render/src/wscript
@@ -16,15 +16,31 @@ def build(bld):
     else:
         sources += ['platform/renderer_common.cpp']
 
-    obj = bld.stlib(features        = 'cxx ddf',
-                    includes        = ['.', '../build/proto'],
-                    proto_gen_py    = True,
-                    protoc_includes = '../proto',
-                    source          = sources,
-                    target          = 'render')
+    # remove the font render plugins
+
+    for x in sources:
+        print(x.name, type(x.name))
+        break
+
+    sources = [x for x in sources if not str(x).endswith('font_renderer_default.cpp')]
+
+    bld.stlib(features        = 'cxx ddf',
+              includes        = ['.', '../build/proto'],
+              proto_gen_py    = True,
+              protoc_includes = '../proto',
+              source          = sources,
+              target          = 'render')
+
+    bld.stlib(features        = 'cxx ddf',
+              includes        = ['.', '../build/proto'],
+              proto_gen_py    = True,
+              protoc_includes = '../proto',
+              source          = ['render/font_renderer_default.cpp'],
+              target          = 'render_font_default')
 
     dmsdk_add_files(bld, '${PREFIX}/sdk/include/dmsdk', 'dmsdk')
 
+    bld.install_files('${PREFIX}/include/render', 'render/font.h')
     bld.install_files('${PREFIX}/include/render', 'render/font_renderer.h')
     bld.install_files('${PREFIX}/include/render', 'render/display_profiles.h')
     bld.install_files('${PREFIX}/include/render', 'render/render.h')
